pyproject.toml
[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[project]
name = "simfast"
version = "1.0.0"
description = "Blazing-fast similarity scores for strings, vectors, points, and sets."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Ali Can Gumusrende", email = "alican.gumusrende@gmail.com" }]
keywords = ["similarity", "distance", "cosine", "levenshtein", "jaccard", "haversine"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Operating System :: OS Independent",
  "Topic :: Software Development :: Libraries",
  "Topic :: Scientific/Engineering :: Information Analysis",
]
dependencies = [
  "numpy>=1.23",
]

[project.optional-dependencies]
fast = [
  "numba>=0.58",
]
dev = [
  "pytest>=8.0",
  "ruff>=0.4",
]

ann-hnsw = [
  "hnswlib>=0.8.0",
]
ann-faiss = [
  "faiss-cpu>=1.7.4",
]
ann = [
  "hnswlib>=0.8.0",
]

[project.urls]
Homepage = "https://pypi.org/project/simfast/"
Repository = "https://github.com/algumusrende/simfast"

[tool.hatch.build.targets.wheel]
packages = ["simfast"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
